Fixing the issue #1
The problem is with github-pages, which enables only safe plugins. So the ones in _plugins folders are not executed. In that case, you have to manually edit the _data/menus.yml file.

For consistency, this change disables unsafe plugins, just like github-pages and does it manually.

Anyway, enabling unsafe plugins will run the hook, but it's modified to skip entries, that are already added manually.